unity-companion-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
chromadb==0.4.22
langchain
langchain-google-genai
python-dotenv==1.0.0
pydantic